Split the state-file format out of the store: format.ts is pure
format.ts owns the on-disk vocabulary (version, page-id grammar) and
parseMap/serializeMap with their replay-through-domain validation (P1) —
no node:* imports, so browsers can consume the format as-is. store.ts
keeps the Node half: atomic writes (P2), page listing, focus requests;
it re-exports the format surface so Node consumers keep one import site.
No behavior change: store tests pass unmodified.
